refactor: 공통 사용되는 Info text 컴포넌트 구조 변경 - #52
